requirements.txt
streamlit
openai
Pillow
pymupdf==1.23.8
pandas
anthropic